fix(runtime): resolve cyclic module dependencies in state action modules
