+++ packages/runtime/src/state/actions/internal/action-types.ts
@@ -1,0 +1,7 @@
+import { ReadOnlyActionTypes } from './read-only-action-types'
+import { ReadWriteActionTypes } from './read-write-action-types'
+
+export const InternalActionTypes = {
+  ...ReadOnlyActionTypes,
+  ...ReadWriteActionTypes,
+} as const
