+++ packages/runtime/src/state/actions/internal/read-only-actions.ts
@@ -16,35 +16,7 @@ import { type DescriptorsByComponentType } from '../../modules/prop-controllers'
 import { type LocaleString, localeStringSchema } from '../../../locale'
 import { type DocumentPayload } from '../../shared-api'
 
-export const ReadOnlyActionTypes = {
-  // TODO: this one should be a read-write action and we should refuse
-  // to fetch resources on the client in read-only mode
-  API_RESOURCE_FULFILLED: 'API_RESOURCE_FULFILLED',
-
-  CREATE_ELEMENT_TREE: 'CREATE_ELEMENT_TREE',
-  DELETE_ELEMENT_TREE: 'DELETE_ELEMENT_TREE',
-
-  REGISTER_COMPONENT: 'REGISTER_COMPONENT',
-  UNREGISTER_COMPONENT: 'UNREGISTER_COMPONENT',
-  REGISTER_COMPONENT_HANDLE: 'REGISTER_COMPONENT_HANDLE',
-  UNREGISTER_COMPONENT_HANDLE: 'UNREGISTER_COMPONENT_HANDLE',
-
-  REGISTER_PROP_CONTROLLERS: 'REGISTER_PROP_CONTROLLERS',
-  UNREGISTER_PROP_CONTROLLERS: 'UNREGISTER_PROP_CONTROLLERS',
-  REGISTER_PROP_CONTROLLERS_HANDLE: 'REGISTER_PROP_CONTROLLERS_HANDLE',
-  UNREGISTER_PROP_CONTROLLERS_HANDLE: 'UNREGISTER_PROP_CONTROLLERS_HANDLE',
-
-  REGISTER_REACT_COMPONENT: 'REGISTER_REACT_COMPONENT',
-  UNREGISTER_REACT_COMPONENT: 'UNREGISTER_REACT_COMPONENT',
-
-  SET_BREAKPOINTS: 'SET_BREAKPOINTS',
-  UPDATE_CLIENT_BREAKPOINT: 'UPDATE_CLIENT_BREAKPOINT',
-
-  SET_LOCALE: 'SET_LOCALE',
-
-  SET_IS_IN_BUILDER: 'SET_IS_IN_BUILDER',
-  SET_IS_READ_ONLY: 'SET_IS_READ_ONLY',
-} as const
+import { ReadOnlyActionTypes } from './read-only-action-types'
 
 type APIResourceFulfilledAction = {
   type: typeof ReadOnlyActionTypes.API_RESOURCE_FULFILLED
