+++ packages/runtime/src/state/actions/internal/read-write-action-types.ts
@@ -1,0 +1,8 @@
+export const ReadWriteActionTypes = {
+  CHANGE_ELEMENT_TREE: 'CHANGE_ELEMENT_TREE',
+  REGISTER_MEASURABLE: 'REGISTER_MEASURABLE',
+  UNREGISTER_MEASURABLE: 'UNREGISTER_MEASURABLE',
+
+  UPDATE_API_CLIENT_CACHE: 'UPDATE_API_CLIENT_CACHE',
+  CLEAR_API_CLIENT_CACHE: 'CLEAR_API_CLIENT_CACHE',
+} as const
