+++ packages/runtime/src/state/shared-api/action-types.ts
@@ -1,0 +1,7 @@
+// bidirectional action types, both the host and the builder can dispatch these actions
+export const SharedActionTypes = {
+  MAKESWIFT_CONNECTION_INIT: 'MAKESWIFT_CONNECTION_INIT',
+
+  REGISTER_DOCUMENT: 'REGISTER_DOCUMENT',
+  UNREGISTER_DOCUMENT: 'UNREGISTER_DOCUMENT',
+} as const
